app/requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
